requirements.txt
numpy
matplotlib
pytest-cov